Typos and moved test
