+++ src/components/cluster/detail/__tests__/NodePoolDropdownMenu.js
@@ -1,0 +1,21 @@
+import '@testing-library/jest-dom/extend-expect';
+import { fireEvent, render } from '@testing-library/react';
+import { ThemeProvider } from 'emotion-theming';
+import React from 'react';
+import theme from 'styles/theme';
+
+// Components
+import NodePoolDropdownMenu from 'cluster/detail/NodePoolDropdownMenu';
+// Not really needed actually. Keeping it by now as an example of simple test without store.
+it('shows the dropdown when the three dots button is clicked', () => {
+  const div = document.createElement('div');
+  const { getByText, getByRole } = render(
+    <ThemeProvider theme={theme}>
+      <NodePoolDropdownMenu render={{ isOpen: true }} />
+    </ThemeProvider>,
+    div
+  );
+  fireEvent.click(getByText('•••'));
+  const menu = getByRole('menu');
+  expect(menu).toBeInTheDocument();
+});
